Avoid Deprecation warning
PHP Deprecated:  substr(): Passing null to parameter #1 ($string) of type string is deprecated in /usr/local/ispconfig/server/plugins-available/bind_plugin.inc.php on line 397

+++ server/plugins-available/bind_plugin.inc.php
@@ -393,7 +393,7 @@ class bind_plugin {
 		//}
 
 		//* Delete old domain file, if domain name has been changed
-		if($data['old']['origin'] != $data['new']['origin']) {
+		if(!empty($data['old']['origin']) && $data['old']['origin'] != $data['new']['origin']) {
 			$filename = $dns_config['bind_zonefiles_dir'].'/' . $this->zone_file_prefix() . str_replace("/", "_", substr($data['old']['origin'], 0, -1));
 
 			if(is_file($filename)) unlink($filename);
